refactor: extract 13 inner classes and localization from CompositeResourceManager
Extract 13 private static final inner classes to InternalStateTypes.java
as package-private top-level classes. Extract localize()/localizeSidekicks()
and SIDEKICK_LABEL_EPILOGUE to CompositeLocalizationDelegate as static
utility methods receiving maps as parameters.

CompositeResourceManager reduced from 649 to 250 lines. Public API unchanged.
All 85 tests pass (29 CRM + 14 InternalStateTypes + 10 LocalizationDelegate
+ 15 TabManager + 17 ViewLifecycle).

Resolves #631
