+++ core/croquet/src/main/java/org/lgna/croquet/CompositeResourceManager.java
@@ -43,414 +43,29 @@
 
 package org.lgna.croquet;
 
-import edu.cmu.cs.dennisc.java.awt.datatransfer.ClipboardUtilities;
 import edu.cmu.cs.dennisc.java.util.Maps;
-import edu.cmu.cs.dennisc.java.util.logging.Logger;
 import org.lgna.croquet.codecs.EnumCodec;
 import org.lgna.croquet.data.ListData;
 import org.lgna.croquet.data.MutableListData;
 import org.lgna.croquet.data.RefreshableListData;
-import org.lgna.croquet.imp.cascade.BlankNode;
 import org.lgna.croquet.preferences.PreferenceBooleanState;
 import org.lgna.croquet.preferences.PreferenceStringState;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.IdentityHashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 
 /**
- * Owns all internal state classes, state maps, factory methods, contains(),
- * and localize() logic extracted from {@link AbstractComposite}.
+ * Owns state maps, factory methods, contains(), and localize() logic
+ * extracted from {@link AbstractComposite}.
+ *
+ * <p>Internal state types live in {@link InternalStateTypes} (same package).
+ * Localization logic is delegated to {@link CompositeLocalizationDelegate}.</p>
  */
 class CompositeResourceManager {
-
-  // ── Private internal state classes ──────────────────────────────────
-
-  private static final class InternalStringValue extends AbstractComposite.AbstractInternalStringValue {
-    private InternalStringValue(AbstractComposite.Key key) {
-      super(UUID.fromString("142b66a2-0b95-42d0-8ea4-a22a79c8ff8c"), key);
-    }
-  }
-
-  private static final class InternalStringState extends StringState {
-    private final AbstractComposite.Key key;
-
-    private InternalStringState(String initialValue, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("ed65869f-8d26-48b1-8240-cf74ba403a2f"), initialValue);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalPreferenceStringState extends PreferenceStringState {
-    private final AbstractComposite.Key key;
-    private final BooleanState isStoringPreferenceDesiredState;
-
-    private InternalPreferenceStringState(String initialValue, AbstractComposite.Key key, BooleanState isStoringPreferenceDesiredState, UUID encryptionId) {
-      super(Application.INHERIT_GROUP, UUID.fromString("ad98acf0-db41-43a6-8619-269a7d8cbc2d"), initialValue, key.getPreferenceKey(), getEncryptionKey(encryptionId != null ? encryptionId.toString() : null));
-      this.key = key;
-      this.isStoringPreferenceDesiredState = isStoringPreferenceDesiredState;
-    }
-
-    @Override
-    protected boolean isStoringPreferenceDesired() {
-      return (this.isStoringPreferenceDesiredState == null) || this.isStoringPreferenceDesiredState.getValue();
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalBooleanState extends BooleanState {
-    private final AbstractComposite.Key key;
-
-    private InternalBooleanState(boolean initialValue, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("5053e40f-9561-41c8-835d-069bd106723c"), initialValue);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalPreferenceBooleanState extends PreferenceBooleanState {
-    private final AbstractComposite.Key key;
-
-    private InternalPreferenceBooleanState(boolean initialValue, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("034f99f7-74ec-4a89-8396-3c187d9684a2"), initialValue, key.getPreferenceKey());
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalSingleSelectListState<T> extends SingleSelectListState<T, ListData<T>> {
-    private final AbstractComposite.Key key;
-
-    private InternalSingleSelectListState(int selectionIndex, ListData<T> data, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("4f0640c9-eceb-4801-a8bb-bf8e282cef0f"), selectionIndex, data);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalImmutableDataSingleSelectListState<T> extends ImmutableDataSingleSelectListState<T> {
-    private final AbstractComposite.Key key;
-
-    private InternalImmutableDataSingleSelectListState(int selectionIndex, ItemCodec<T> codec, T[] values, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("091d5251-d278-4eb1-8214-a27c154f5378"), selectionIndex, codec, values);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalRefreshableDataSingleSelectListState<T> extends RefreshableDataSingleSelectListState<T> {
-    private final AbstractComposite.Key key;
-
-    private InternalRefreshableDataSingleSelectListState(int selectionIndex, RefreshableListData<T> data, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("4d7ef91c-a8ae-4b17-9d8a-91ffac4ba12e"), selectionIndex, data);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalMutableDataSingleSelectListState<T> extends MutableDataSingleSelectListState<T> {
-    private final AbstractComposite.Key key;
-
-    private InternalMutableDataSingleSelectListState(int selectionIndex, MutableListData<T> data, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("6cc16988-0fc8-476b-9026-b19fd15748ea"), selectionIndex, data);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalTabState<T extends SimpleTabComposite<?>> extends SimpleTabState<T> {
-    private final AbstractComposite.Key key;
-
-    public InternalTabState(int selectionIndex, Class<T> cls, T[] values, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("bea99c2f-45ad-40a8-a99c-9c125a72f0be"), selectionIndex, cls, values);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalBoundedIntegerState extends BoundedIntegerState {
-    private final AbstractComposite.Key key;
-
-    private InternalBoundedIntegerState(BoundedIntegerState.Details details, AbstractComposite.Key key) {
-      super(details);
-      this.key = key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalBoundedDoubleState extends BoundedDoubleState {
-    private final AbstractComposite.Key key;
-
-    private InternalBoundedDoubleState(BoundedDoubleState.Details details, AbstractComposite.Key key) {
-      super(details);
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
-
-  private static final class InternalCascadeWithInternalBlank<T> extends CascadeWithInternalBlank<T> {
-    private final AbstractComposite.CascadeCustomizer<T> customizer;
-    private final AbstractComposite.Key key;
-
-    private InternalCascadeWithInternalBlank(AbstractComposite.CascadeCustomizer<T> customizer, Class<T> componentType, AbstractComposite.Key key) {
-      super(Application.INHERIT_GROUP, UUID.fromString("165e65a4-fd9b-4a09-921d-ecc3cc808de0"), componentType);
-      this.customizer = customizer;
-      this.key = key;
-    }
-
-    public AbstractComposite.Key getKey() {
-      return this.key;
-    }
-
-    @Override
-    protected Class<? extends Element> getClassUsedForLocalization() {
-      return this.key.getComposite().getClass();
-    }
-
-    @Override
-    protected String getSubKeyForLocalization() {
-      return this.key.getLocalizationKey();
-    }
-
-    @Override
-    protected org.lgna.croquet.edits.Edit createEdit(org.lgna.croquet.history.UserActivity userActivity, T[] values) {
-      return this.customizer.createEdit(values);
-    }
-
-    @Override
-    protected List<CascadeBlankChild> updateBlankChildren(List<CascadeBlankChild> rv, BlankNode<T> blankNode) {
-      this.customizer.appendBlankChildren(rv, blankNode);
-      return rv;
-    }
-
-    @Override
-    protected void appendRepr(StringBuilder sb) {
-      super.appendRepr(sb);
-      sb.append(";key=");
-      sb.append(this.key);
-    }
-  }
 
   // ── State maps ──────────────────────────────────────────────────────
 
@@ -602,48 +217,34 @@ class CompositeResourceManager {
     return this.containsIndex.contains(model);
   }
 
-  // ── localize() ──────────────────────────────────────────────────────
+  // ── Map accessors for localization delegate ──────────────────────────
 
-  private static final String SIDEKICK_LABEL_EPILOGUE = ".sidekickLabel";
-
-  @SuppressWarnings("unchecked")
-  private void localizeSidekicks(AbstractComposite<?> composite, Map<AbstractComposite.Key, ? extends CompletionModel>... maps) {
-    for (Map<AbstractComposite.Key, ? extends CompletionModel> map : maps) {
-      for (Map.Entry<AbstractComposite.Key, ? extends CompletionModel> entry : map.entrySet()) {
-        AbstractComposite.Key key = entry.getKey();
-        CompletionModel model = entry.getValue();
-        String text = composite.findLocalizedText(key.getLocalizationKey() + SIDEKICK_LABEL_EPILOGUE);
-        if (text != null) {
-          StringValue sidekickLabel = model.getSidekickLabel();
-          text = composite.modifyLocalizedText(sidekickLabel, text);
-          sidekickLabel.setText(text);
-        } else {
-          if (model.hasSidekickLabel()) {
-            Class<?> cls = composite.getClassUsedForLocalization();
-            String localizationKey = cls.getSimpleName() + "." + key.getLocalizationKey() + SIDEKICK_LABEL_EPILOGUE;
-            Logger.errln();
-            Logger.errln("WARNING: could not find localization for sidekick label");
-            Logger.errln("looking for:");
-            Logger.errln();
-            Logger.errln("   ", localizationKey);
-            Logger.errln();
-            Logger.errln("in croquet.properties file in package:", cls.getPackage().getName());
-            Logger.errln();
-            Logger.errln(localizationKey, "has been copied to the clipboard for your convenience.");
-            Logger.errln("if this does not solve your problem please feel free to ask dennis for help.");
-            ClipboardUtilities.setClipboardContents(localizationKey);
-          }
-        }
-      }
-    }
+  Map<AbstractComposite.Key, AbstractComposite.AbstractInternalStringValue> getStringValueMap() {
+    return this.mapKeyToStringValue;
   }
 
   @SuppressWarnings("unchecked")
+  Map<AbstractComposite.Key, ? extends CompletionModel>[] getSidekickMaps() {
+    return new Map[] {
+        this.mapKeyToActionOperation,
+        this.mapKeyToBooleanState,
+        this.mapKeyToPreferenceBooleanState,
+        this.mapKeyToBoundedDoubleState,
+        this.mapKeyToBoundedIntegerState,
+        this.mapKeyToCascade,
+        this.mapKeyToItemState,
+        this.mapKeyToImmutableSingleSelectListState,
+        this.mapKeyToRefreshableSingleSelectListState,
+        this.mapKeyToMutableSingleSelectListState,
+        this.mapKeyToTabState,
+        this.mapKeyToPreferenceStringState,
+        this.mapKeyToStringState,
+    };
+  }
+
+  // ── localize() ──────────────────────────────────────────────────────
+
   void localize(AbstractComposite<?> composite) {
-    for (Map.Entry<AbstractComposite.Key, AbstractComposite.AbstractInternalStringValue> entry : this.mapKeyToStringValue.entrySet()) {
-      AbstractComposite.AbstractInternalStringValue stringValue = entry.getValue();
-      stringValue.setText(composite.modifyLocalizedText(stringValue, composite.findLocalizedText(entry.getKey().getLocalizationKey())));
-    }
-    this.localizeSidekicks(composite, this.mapKeyToActionOperation, this.mapKeyToBooleanState, this.mapKeyToPreferenceBooleanState, this.mapKeyToBoundedDoubleState, this.mapKeyToBoundedIntegerState, this.mapKeyToCascade, this.mapKeyToItemState, this.mapKeyToImmutableSingleSelectListState, this.mapKeyToRefreshableSingleSelectListState, this.mapKeyToMutableSingleSelectListState, this.mapKeyToTabState, this.mapKeyToPreferenceStringState, this.mapKeyToStringState);
+    CompositeLocalizationDelegate.localize(composite, this.mapKeyToStringValue, this.getSidekickMaps());
   }
 }
